feat(RegistrationHtml)first version of the registration endpoint and HTML page added
