+++ src/main/java/com/springles/tickets/controllers/UserController.java
@@ -1,26 +1,37 @@
 package com.springles.tickets.controllers;
 
 import com.springles.tickets.models.ApplicationUser;
+import com.springles.tickets.models.Doctor;
 import com.springles.tickets.repositories.ApplicationUserRepository;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
-@RestController
-@RequestMapping("/users")
+@Controller
+@RequestMapping("/register")
 public class UserController {
 
   private ApplicationUserRepository applicationUserRepository;
   private BCryptPasswordEncoder bCryptPasswordEncoder;
 
   public UserController(ApplicationUserRepository applicationUserRepository,
-      BCryptPasswordEncoder bCryptPasswordEncoder) {
+                        BCryptPasswordEncoder bCryptPasswordEncoder) {
     this.applicationUserRepository = applicationUserRepository;
     this.bCryptPasswordEncoder = bCryptPasswordEncoder;
   }
 
-  @PostMapping("/sign-up")
-  public void signUp(@RequestBody ApplicationUser user) {
-    user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
-    applicationUserRepository.save(user);
+  @GetMapping("/")
+  public String showRegistrationForm(@ModelAttribute(value = "new_user") ApplicationUser newUser,
+                                     Model model){
+    model.addAttribute("new_user", newUser);
+    return "new_user_form";
+  }
+
+  @PostMapping("/")
+  public String signUp(@ModelAttribute(value = "new_user") ApplicationUser newUser) {
+    newUser.setPassword(bCryptPasswordEncoder.encode(newUser.getPassword()));
+    applicationUserRepository.save(newUser);
+    return "redirect:/welcome";
   }
 }
