+++ src/main/java/com/springles/tickets/controllers/UserRestController.java
@@ -1,0 +1,27 @@
+package com.springles.tickets.controllers;
+
+import com.springles.tickets.models.ApplicationUser;
+import com.springles.tickets.repositories.ApplicationUserRepository;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/users")
+public class UserRestController {
+
+  private ApplicationUserRepository applicationUserRepository;
+  private BCryptPasswordEncoder bCryptPasswordEncoder;
+
+  public UserRestController(ApplicationUserRepository applicationUserRepository,
+                            BCryptPasswordEncoder bCryptPasswordEncoder) {
+    this.applicationUserRepository = applicationUserRepository;
+    this.bCryptPasswordEncoder = bCryptPasswordEncoder;
+  }
+
+  @PostMapping("/sign-up")
+  public void signUp(@RequestBody ApplicationUser user) {
+    user.setPassword(bCryptPasswordEncoder.encode(user.getPassword()));
+    applicationUserRepository.save(user);
+  }
+}
+
